moved class finder to common
