--- tests/Automatic/Fixture/Finder/FooTrait.php
@@ -1,7 +1,0 @@
-<?php
-declare(strict_types=1);
-namespace Narrowspark\Automatic\Test\Fixture\Finder;
-
-trait FooTrait
-{
-}
