+++ tests/Common/Fixture/Finder/DummyClass.php
@@ -1,18 +1,18 @@
 <?php
 declare(strict_types=1);
-namespace Narrowspark\Automatic\Test\Fixture\Finder;
+namespace Narrowspark\Automatic\Common\Test\Fixture\Finder;
 
 use Narrowspark\Automatic\Common\Contract\Configurator as ConfiguratorContract;
 use Narrowspark\Automatic\Common\Contract\Package as PackageContract;
 
-class DummyClassTwo implements ConfiguratorContract
+class DummyClass implements ConfiguratorContract
 {
     /**
      * {@inheritdoc}
      */
     public static function getName(): string
     {
-        return 'two';
+        return 'class';
     }
 
     /**
