+++ tests/Common/Fixture/Finder/FooTrait.php
@@ -1,0 +1,7 @@
+<?php
+declare(strict_types=1);
+namespace Narrowspark\Automatic\Common\Test\Fixture\Finder;
+
+trait FooTrait
+{
+}
